+++ src/main/java/com/backend/onlinecvproject/repositories/CandidateRepository.java
@@ -3,5 +3,5 @@ package com.backend.onlinecvproject.repositories;
 import com.backend.onlinecvproject.entities.Candidate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CandidateRepository extends JpaRepository<Candidate, Long> {
+public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
 }
