+++ src/main/java/com/backend/onlinecvproject/repositories/WorkExperienceRepository.java
@@ -3,5 +3,5 @@ package com.backend.onlinecvproject.repositories;
 import com.backend.onlinecvproject.entities.WorkExperience;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkExperienceRepository extends JpaRepository<WorkExperience, Long> {
+public interface WorkExperienceRepository extends JpaRepository<WorkExperience, Integer> {
 }
